move AutoRelease class to its own file
